+++ CHANGELOG.md
@@ -91,6 +91,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
  - `Cursor::saveState()` and `Cursor::restoreState()` now use `CursorState` objects instead of arrays
  - `NodeWalker::next()` now enters, traverses any children, and leaves all elements which may have children (basically all blocks plus any inlines with children). Previously, it only did this for elements explicitly marked as "containers".
  - `InvalidOptionException` now extends from `UnexpectedValueException`
+ - The line numbers (keys) returned via `MarkdownInput::getLines()` now start at 1 instead of 0
 
 ### Removed